requirements.txt
streamlit>=1.45,<2
pandas>=2.2,<3
zstandard>=0.22,<1
pycryptodome>=3.21,<4
